duckdb/pyproject.toml
[tool.poetry]
name = "duckdb-pl-pd"
version = "0.1.0"
description = ""
authors = ["mehd-io <mehdi@mehd.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
duckdb = "^0.8.0"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
